+++ test/WaterfallRenderContext.test.mjs
@@ -1,8 +1,9 @@
-import { strictEqual } from 'assert';
+import { ok, strictEqual } from 'assert';
 import { useContext } from 'react';
 import { renderToStaticMarkup } from 'react-dom/server.js';
 import { jsx } from 'react/jsx-runtime.js';
 import WaterfallRenderContext from '../public/WaterfallRenderContext.js';
+import getBundleSize from './getBundleSize.mjs';
 
 export default (tests) => {
   tests.add('`WaterfallRenderContext` used as a React context.', () => {
@@ -18,5 +19,13 @@ export default (tests) => {
       ),
       contextValue
     );
+  });
+
+  tests.add('`WaterfallRenderContext` bundle size.', async () => {
+    const kB = await getBundleSize(
+      new URL('../public/WaterfallRenderContext.js', import.meta.url)
+    );
+
+    ok(kB < 0.5);
   });
 };